backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pandas
matplotlib
reportlab
python-dateutil
